move NodeID to node package
